requirements.txt
keyboard
pyautogui
Pillow
